dispatchd: Fix bind(): Address already in use errors
Caused by socket being placed into a wait state by the OS
when this end is closed first (see address-in-use.pdf).
Make sure the client closes its end of the socket first.
Signal the client by sending EOT, then wait for the client to
close by issuing a read() and checking for EOF.